In addition to mouse events, keyboard events will propagate down the GUI tree

+++ GameEngine/GameEngine/Callbacks.cpp
@@ -9,7 +9,7 @@ Callbacks::Callbacks() {}
 Callbacks::~Callbacks() {}
 
 void Callbacks::CursorPosCallback(GLFWwindow * window, double xpos, double ypos) {
-	Input::feedMousePosition(Screen::fromScreenToUISpace(Vector2f(xpos, ypos)));
+	Input::feedMousePosition(Screen::fromScreenToUISpace(Vector2f((float)xpos, (float)ypos)));
 }
 
 void Callbacks::KeyCallback(GLFWwindow * window, int key, int scancode, int action, int mods) {
@@ -23,14 +23,14 @@ void Callbacks::MouseCallback(GLFWwindow * window, int button, int action, int m
 void Callbacks::WindowSizeCallback(GLFWwindow * window, int width, int height) {
 	if (Window::isInitialized()) {
 		glViewport(0, 0, width, height);
-		Screen::updateRes(width, height);
+		Screen::updateRes((float)width, (float)height);
 		FramebufferManager::updateSizes(width, height);
 		Window::setSize(width, height);
 	}
 }
 
 void Callbacks::ScrollCallback(GLFWwindow * window, double xoffset, double yoffset) {
-	Screen::updateScroll(yoffset / 200.0f);
+	Screen::updateScroll((float)yoffset / 200.0f);
 }
 
 void Callbacks::WindowPosCallback(GLFWwindow * window, int x, int y)
